added a ouput_HR folder where chris can store the output images,  edited dataloader to directly take data from the script files of Trainer and LoadData,
Chris, store the output images from the model to output_HR file.

The dataloader.ipynb isn't working on my local computer. Specifically the error is that It is not getting any image even though there are 20 images. I think its not a big deal and you'll be able to correct it if you want to. But this is not a priority for the report.